## Why
The current secondary evidences are taken from the ingame *Journal*.
These often do not really describe what the actual secondary evidences are.

With the introduction of the ***Nightmare*** mode the secondary evidences became crucial which is the reason for the update.

## What
`data.js` got a reworked list of all secondary evidences with the current state of our knowledge about them.
The source for the new list is mainly the [Phasmophobia Wiki](https://phasmophobia.fandom.com/wiki/Main_Page), [patchnotes 0.4.0](https://store.steampowered.com/news/app/739630?emclan=103582791466524086&emgid=4624626247919023465), [patchnotes 0.4.1](https://store.steampowered.com/news/app/739630/view/3060752450167367025) and the latest [Dev FAQ video uploaded by Insym](https://www.youtube.com/watch?v=orMdTJvdOVM) where CJ directly answers questions about the new ghosts.

## Testing
I have no npm to test it and I only have edited the json so please test for me 🙏 - apologies.