@un0btanium 2021-11-02
Wow this is awesome. I currently have very little time and I have not actively been following Phasmophobia. Can someone else doublecheck if this is all correct?