@Redforce04 2021-11-02
@BotiKis I updated a few minor things and created a pr on your master branch. Once you merge those, it should be ready to go. Most of it is minor json things to make it parse properly in online parsers, and then I added your name in the about page.