@Redforce04 2021-11-04
I will double check the info although most of it looks factual.