@Redforce04 2021-11-04
This will be a long comment but I will cross check the new evidence with the wiki

- [ ] Chance to throw objects at great speeds
- [ ] More ghost activity and events when there are more people are nearby
- [ ] Smudge sticks stop attacks for 3 min instead of 1.5
- [ ] Stepping in salt leaves no UV footprints
- [ ] Can teleport to a random player when not hunting
- [ ] Looking at the ghost drops sanity by 0.4% each second instead of 0.2%
- [ ] Taking a photo makes the ghost temporarily disappear
- [ ] Blinks slower during hunts (1s to 2s instead of 0.3s to 1s)
- [ ] Throws multiple objects at once and reduces sanity by the number of thrown items times 2
- [ ] Targets and hunts one player until killed (If the player is not in the building, it does a regular hunt)
- [ ] Can start a hunt early regardless the avarage sanity if it sees the choosen player long enough
- [ ] The Crucifixs effective radius is 5m instead of 3m
- [ ] Only choosen player can finish the 'Repel a ghost while it is chasing' objective
- [ ] Can lower the sanity of all players within a certain range by 25%
- [ ] During a hunt it travels faster when chasing a target until it comes close
- [ ] Cannot use abilities when the locations power is turned off
- [ ] Can hunt at 60 % avg sanity when lights are turned off
- [ ] Cannot turn on lights
- [ ] Very fast when chasing a player during a hunt (2x normal speed)
- [ ] Very slow when not chasing anyone during a hunt (1/2 of normal speed)
- [ ] Less chance of ghost activity and ghost events when more than 1 player is nearby
- [ ] Lower chance of initiating a hunt when more than 1 player is nearby
- [ ] Can initiate a hunt at an avarage sanity of 70%
- [ ] Successful answers from the ouja board lowers sanity by 40% instead of 50%
- [ ] Ghost event drops sanity by 0.4% per second instead of 0.2%
- [ ] Smudging will make it roam less on average for 1.5 min
- [ ] Can initiate a hunt at an avarage sanity of 80% if players are talking nearby
- [ ] During a hunt it only can hear sounds in a radius of 2m
- [ ] Walking speed is affected by room temperature - hot 1.8 m/s, cold 1 m/s
- [ ] D.O.T.S. only visible through a video cam and no players in the room
- [ ] Less likely to roam when not hunting
- [ ] Always gives the D.O.T.S. evidence on nightmare difficulty
- [ ] More sounds on the parabolic mic
- [ ] During a hunt footstep sounds can only be heard when 10m or closer (similar to electronic interference)
- [ ] Can initiate a hunt when any flames are extinguished regardless of average sanity
- [ ] Chance for the ability increases everytime a player is killed
- [ ] Main twin primary stays in the ghost room, makes freezing and spirit box
- [ ] Secondary twin wanders around interacting with the environment
- [ ] Secondary twin cannot interact with motion sensors, spirit box or make freezing
- [ ] Either Twin can initiate a Hunt but not simultaneously
- [ ] While hunting it has an increased movement speed when it touches electronic equipment
- [ ] It doesn't matter if the equipment lies on the floor or held by a player
- [ ] Abilities only affected by player equipment and not location electrics like TVs or lights
- [ ] Can initiate a hunt at an avarage sanity of 70% if there is electrical equipment nearby
- [ ] Reduced chance to leave fingerprints
- [ ] Low chance to leave fingerprints with 6 fingers
- [ ] Fingerprints disappear earlier (60s) than other ghosts (120s)
- [ ] Always gives the Fingerprints evidence on nightmare difficulty