@Redforce04 2021-11-05
I have confirmed most of these but I have some suggestions.
## We may want to change the jinn's evidence so they mention that power must be on IE:
#### Remove this - `Cannot use abilities when the locations power is turned off` 
#### Change the others:
 - `Has the ability to lower the sanity of all players within a certain range by 25% when power is on`
 - `During a hunt it, while the power is on, the ghost travels faster when chasing a target until it gets close.`

## Myling: 
` "More sounds on parabolic mic" ` isn't very descriptive or helpful. Maybe put something like `"louder sounds on the parabolic mic are common"` or `"Sounds on the parabolic mic are frequent"` to make it a bit more helpful 

## Raiju: 
Hunts at 65% not 70% according to wiki.

 ## Other:
Found no mention of ghost evidences guaranteed to show up on nightmare mode.
May want to re-write these so it flows better. IE: 
`"Low chance to leave fingerprints with 6 fingers"` -> `"A 6-finger fingerprint was found."`
I can rewrite these and suggest.