@Redforce04 2021-11-05
I will edit this comment with all of the re-written suggestions, then I will create a pr and it should be ready to merge.