@Redforce04 2021-11-05
Before I make the pr I wanna see if anyone has suggestions for this pull or wants me to re-write some of the evidence above. It is all factually confirmed currently.