@BotiKis 2021-11-15
@un0btanium Sorry for the dumb question, but how/when does the [page](https://un0btanium.github.io/phasmophobia-cheatsheet/) update?