@un0btanium 2021-11-15
not dumb at all. i actually forgot to deploy it to github pages. sorry about that. thanks for reminding me. if possible i probably should set up an autodeploy github script.